marketplace_builder_shared

Module coordinator

Modules

tiered_view_map

This module contains an optimized implementation of a composite key two-tier map, where the first key is a view number

Structs

BuilderStateCoordinator

A coordinator managing the lifecycle of BuilderStates.

Enums

BuilderStateLookup

Result of looking up a builder state by ID.

Type Aliases

BuilderStateMap 🔒

ProposalMap 🔒