marketplace_builder_shared::coordinator
Module tiered_view_map

This module contains an optimized implementation of a composite key two-tier map, where the first key is a view number

Structs

TieredViewMap

A map from ViewCompositeKey to arbitrary value, implemented as a tiered map with the first tier being BTreeMap keyed by view number of ViewCompositeKey and the second [NEMap] keyed by subkey of ViewCompositeKey.

Traits

ViewCompositeKey

A two-component key, of which one component is ConsensusTime