marketplace_builder_shared
Module testing

Modules

consensus

This module defines types used when simulating consensus in tests

constants

Shared constants for tests

generation

mock

A collection of generator functions for mock data used in tests

validation

Structs

TransactionPayload

Transaction payload used in testing to track various properties of individual transactions

Functions

run_test

Run a builder test. Thin wrapper around TestRunner::run_test, injecting transaction generation and validation tasks based on validation_config and transaction_generation_config respectively.