marketplace_builder_shared::testing
Module constants

Shared constants for tests

Constants

TEST_API_TIMEOUT

Governs API timeout for builder API. This is an arbitrary default value for testing.

TEST_BASE_FEE

Governs fee per byte used by builders. This is an arbitrary default value for testing.

TEST_CHANNEL_BUFFER_SIZE

Governs the buffer size used for the test channels. All of the channels created need a capacity. The concrete capacity isn’t specifically bounded in tests, so it is set to an arbitrary value.

TEST_INCLUDED_TX_GC_PERIOD

Governs the included transaction GC period used in tests. This is an arbitrary default value for testing.

TEST_MAXIMIZE_TX_CAPTURE_TIMEOUT

Governs timeout when waiting to fill transaction queue on incoming bundle/block request. This is an arbitrary default value for testing.

TEST_MAX_BLOCK_SIZE_INCREMENT_PERIOD

Controls period between optimistic increments of maximum block size limit. This is an arbitrary default value for testing.

TEST_MAX_TX_NUM

TEST_NUM_CONSENSUS_RETRIES

Controls the number of attempts that the simulated consensus will perform when an error is returned from the Builder when asking for available blocks.

TEST_NUM_NODES_IN_VID_COMPUTATION

Controls the number of nodes that are used in the VID computation for the tests. This is an arbitrary default value for testing.

TEST_PROTOCOL_MAX_BLOCK_SIZE

Controls the maximum block size of this protocol instance. This is an arbitrary default value for testing.

TEST_TX_STATUS_CACHE_CAPACITY

Governs the target space used by the mapping from txn to its status. This is expressed as a target number of transactions. This is an arbitrary default value for testing.