marketplace_builder_shared::testing

Module generation

Structs

TransactionGenerationConfig

Configuration for transaction generation task

TransactionGenerationTask

Enums

GenerationStrategy

Strategy for generating transactions

SubmissionEndpoint

Strategy for transaction submission Public will use public mempool, submitting to a node with given index Private will use private mempool

Constants

FLOOD_INCREMENT 🔒