marketplace_builder_shared::testing

Module mock

A collection of generator functions for mock data used in tests

Functions

builder_state

decide_leaf_chain

decide_leaf_chain_with_transactions

parent_references

Generate references for given view number with random commitments for use in testing code

proposals

Create mock pair of DA and Quorum proposals

proposals_with_transactions

Create mock pair of DA and Quorum proposals with given transactions

transaction